HOOOO SHIT. I don't know how I was able to do this, but we get cover art!!!!!
Okay, so... attempt to explain this. Files with cover art have a key "covr" (I've only tested AACs so far), we take that data, convert it into a base64 string and then shove it into the WebView that we created earlier. It's probably extremely brittle right now and the completely opposite of DRY, but it works!