Handle files with absolutely no metadata.
I made a lot of assumptions with the way we were handling metadata—such as expecting it to always be there. We now check to see if any of the metadata returns `nil` before assigning the properties to labels. Also, we default to the file's name first and overwrite that if the correct title metadata is supplied.

Thanks goes to @iceymoon's JPLOP forum buddies for supplying a file that alerted me to this problem.