Successfully retrieving metadata from iTunes drops.
Talk about not documented.

When you drag from iTunes to an application using AVAsset, there are
restrictions placed on where the metadata can be fetched from. Even
though we're fetching a file URL, the metadata still comes from iTunes
and not from inside the file. So we have to lighten the restriction on
AVURLAsset so that it can grab the data from iTunes.

Really weird.